Communication issue when retrieving information about a PID mode
To address this issue, we propose implementing two improvements. Firstly, we will increase the timeout value for the HTTP client used in the PID mode
information retrieval process.

Additionally, we will introduce a caching mechanism to store the responses obtained from the PID mode information retrieval.
This caching strategy will eliminate the need for redundant requests when the same PID information is requested multiple times within a short period.